Fixing memcpy, it seems to add 16 rather than 1 on toSend + 1

+++ src/quaternion.h
@@ -1,10 +1,10 @@
 #include <math.h>
 struct Vec4{
-	double v[4];
+	float v[4];
 };
 
 struct Vec3{
-	double v[3];
+	float v[3];
 };
 
 Vec3 Quat2RPY(Vec4 q){
